+++ client/src/app/routes/todos/todolist/todolist.component.ts
@@ -1,8 +1,9 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, EventEmitter, Input, Output, SimpleChanges } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { User, ToDoTask } from '../../../shared/models/auth';
 import { AuthService } from '../../../services/auth.service';
 import { TodosService } from '../../../services/todos.service';
+import { ConfirmBoxInitializer } from '@costlydeveloper/ngx-awesome-popup';
 
 @Component({
   selector: 'app-todolist',
@@ -15,24 +16,28 @@ export class TodolistComponent {
 
 
 @Input() todoTask: ToDoTask | undefined = undefined;
+@Output() updateTask = new EventEmitter<{ task: ToDoTask }>();
+@Output() deleteTask = new EventEmitter<{ task: ToDoTask }>();
+
 
 constructor(
     private authService: AuthService,
     private todosService: TodosService
   ) {}
 
-   updateStatus(task : ToDoTask){
-
-    this.todosService.updateTask(task).subscribe({
-      next: (res) => {
-        console.log(res);
-        //this.loadTodos();
-      },
-      error: (res) => {
-        console.log(res.error.message);
-      },
-    });
-
+   onTaskClick(task : ToDoTask){
+    this.updateTask.emit( { task: task })
   }
+
+  onDeleteClick(task : ToDoTask){
+   this.deleteTask.emit( { task: task })
+  }
+  
+
+  
+  // ngOnChanges(changes: SimpleChanges) {
+  //   console.log("changes here : "+ changes['todoTask'])
+
+  // }
 
 }