+++ client/src/app/services/todos.service.ts
@@ -2,7 +2,8 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, map } from 'rxjs';
 import { environment } from '../../environments/environment.development';
-import { EditTaskResponse, ToDoTask } from '../shared/models/auth';
+import { AddNewTaskModel, DeleteTaskResponse, EditTaskResponse, ToDoTask } from '../shared/models/auth';
+import { TmplAstForLoopBlock } from '@angular/compiler';
 
 @Injectable({
   providedIn: 'root',
@@ -17,7 +18,6 @@ export class TodosService {
       'Content-Type': 'application/json',
       Authorization: 'Bearer ' + localStorage.getItem('token'),
     });
-    console.log(reqHeader);
     return this.http.get<ToDoTask[]>(this.baseUrl + 'tasks/' + userId, {
       headers: reqHeader,
     });
@@ -40,4 +40,34 @@ export class TodosService {
       }),
     );
   }
+  
+  deleteTask(task : ToDoTask): Observable<boolean> {
+    
+    const reqHeader = new HttpHeaders({
+      'Content-Type': 'application/json',
+      Authorization: 'Bearer ' + localStorage.getItem('token'),
+    });
+    return this.http
+    .delete<DeleteTaskResponse>(this.baseUrl + 'Tasks/'+task.id,  {
+      headers: reqHeader,
+    })
+    .pipe(
+      map((res) => {
+        return res.status;
+      }),
+    );
+  }
+
+  addNewTask( task : AddNewTaskModel) : Observable<ToDoTask> {
+
+    const reqHeader = new HttpHeaders({
+      'Content-Type': 'application/json',
+      Authorization: 'Bearer ' + localStorage.getItem('token'),
+    });
+    return this.http.post<ToDoTask>(this.baseUrl + 'tasks', task , {
+      headers: reqHeader,
+    });
+
+  }
+  
 }